refator: useQuery 및 useinfiniteQuery 커스텀 훅으로 수정
