+++ src/component/Search/SearchList.tsx
@@ -11,8 +11,7 @@ import {
 import { useNavigate } from 'react-router-dom'
 import { IPlayList } from '@/types/playList'
 import { useDebounce } from '@/hooks'
-import { useQuery } from '@tanstack/react-query'
-import { getSearchPlayLists } from '@/apis/search/playList'
+import useFetchSearchPlayList from '@/hooks/useFetchSearchList'
 
 // 검색한 내용을 담아냄
 const SearchList = () => {
@@ -39,13 +38,7 @@ const SearchList = () => {
     navigate(`/search?search=${searchValue}`, { state: searchValue }) // navigate로 검색어와 함께 이동
   }
 
-  const { data: searchs, isError } = useQuery<IPlayList[]>({
-    queryKey: ['searchPlayList', debouncedTitle],
-    queryFn: () => getSearchPlayLists(debouncedTitle),
-    enabled: !!debouncedTitle,
-    staleTime: 1000 * 60
-  })
-
+  const { data: searchs, isError } = useFetchSearchPlayList(debouncedTitle)
   if (isError) throw new Error('에러가 발생했습니다.')
 
   return (
