+++ src/hooks/useFetchSearchList.ts
@@ -1,0 +1,14 @@
+import { useQuery } from '@tanstack/react-query'
+import { IPlayList } from '@/types/playList'
+import { getSearchPlayLists } from '@/apis/search/playList'
+
+const useFetchPlayList = (debouncedTitle: string) => {
+  return useQuery<IPlayList[]>({
+    queryKey: ['searchPlayList', debouncedTitle],
+    queryFn: () => getSearchPlayLists(debouncedTitle),
+    enabled: !!debouncedTitle,
+    staleTime: 1000 * 60
+  })
+}
+
+export default useFetchPlayList
